+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 ### Changed
 
 - ♿️(frontend) Focus main container after navigation #1854
+- 🚸(backend) sort user search results by proximity with the active user #1802 
 
 
 ### Fixed